Start re-introducing modified vector as a separate opt-in feature Opt-in for future-proofing
