+++ src/system/stlport/README
@@ -7,6 +7,7 @@ Modifications are made as needed to suit the decompilation project, including bu
 - Only the `stlport` directory has been included from the original package, as STLport iostreams/locale/etc. code are not necessary for this project.
 - Additional compiler configuration headers [`config/stl_mwbroadway.h`](config/stl_mwbroadway.h) and [`config/stl_mwgecko.h`](config/stl_mwgecko.h) have been added to support MWCC for GC/Wii.
 - A custom allocator is used as the default allocator instead of `std::allocator`.
+- A modified version of `vector` has been introduced which accepts a third template parameter for the integer type used for size/capacity, based on symbol/RTTI names and decompiled code.
 
 Any other important modifications or hacks can be found by searching for `// decomp hack:` inside this folder.
